Create README.md
Who is of people at risk for a stroke?
According to some research and studies, stroke is one of the leading causes of death worldwide with 11%. So that we can control some of the risk factors for stroke. I will use this project to find out the risks that put patients at risk of stroke based on certain characteristics.
like: age,gender, hypertension, heart_disease
, ever_married,  work_type,  Residence_type,  avg_glucose_level,  bmi,  smoking_status. 
i was used  stroke feature as lable.

by using classification model , aplly on the diffrent models and comparing between them.
1- logistic regression classifer
2-KNN classifer
3-Random Forest classifer
after modeling i have seen the best model for accuracy is Random Forest classifer, then KNN classiferans the last one is logistic regression classifer.